First commit in the redesign endeavour. May not pan out, but goal is to get a lot of the code out of the main for loop into functions. May help with games and other things.
