+++ src/bot.go
@@ -5,11 +5,9 @@ import (
 	"database/sql"
 	"fmt"
 	"log"
-	"math/rand"
 	"net"
 	"net/textproto"
 	"os"
-	"strconv"
 	"strings"
 	"time"
 
@@ -409,10 +407,6 @@ func main() {
 	reader := bufio.NewReader(irc.conn)
 	proto := textproto.NewReader(reader)
 
-	currenttime := time.Now()
-	datestring := currenttime.String()
-	datesplit := strings.Split(datestring, " ")
-
 	// For more information on this goroutine, look at the server.go file.
 	if irc.WebAppGUIEnabled == true {
 		go ServerMain()
@@ -471,153 +465,11 @@ func main() {
 				fmt.Printf(username + ": " + usermessage + "\n")
 			}
 
-			// Check if user set MakeLog in config.toml to true, if so, run
-			if irc.MakeLog == true {
-				// Use current date to mark which day the chat log is for
-				loglocation := "logs/chat/" + datesplit[0] + ".txt"
-				logmessage := (username + ": " + usermessage + "\n")
-				WriteToLog(loglocation, logmessage)
-			}
-
-			// Check if user set CheckLongMessageCap in config.toml to true, if so, run
-			if irc.CheckLongMessageCap == true {
-				if len(usermessage) > irc.LongMessageCap {
-					fmt.Println("Very long message detected.")
-					PurgeUser(irc, username)
-					BotSendMsg(irc.conn, irc.ChannelName, "@"+username+" please shorten your message", irc.BotName)
-
-				}
-			}
-
-			// For each value in LinkChecks array in config.toml, check whether to purge user or not.
-			for _, v := range irc.LinkChecks {
-				if strings.Contains(usermessage, v) {
-					if irc.PurgeForLinks == true {
-
-						// Check if user is in the permitted slice
-						userCheck := UserInSlice(username, permUsers)
-						// If user is a moderator / broadcaster, just let them post link
-						if CheckUserStatus(line, "moderator", irc) == "true" || CheckUserStatus(line, "broadcaster", irc) == "true" {
-							fmt.Println("Link permitted.")
-						} else if userCheck == true { // If not a moderator / broadcaster, but is in the permitted slice, let them post link then remove them
-							position := GetSlicePosition(username, permUsers)
-							permUsers = RemoveFromSlice(position, permUsers)
-						} else { // If none of the above is true, purge user
-							PurgeUser(irc, username)
-						}
-					}
-				}
-			}
-
-			// Check if a user is moderator or broadcaster before checking conditions for multiple commands.
 			if CheckUserStatus(line, "moderator", irc) == "true" || CheckUserStatus(line, "broadcaster", irc) == "true" {
-				if strings.Contains(usermessage, "!editcom") || strings.Contains(usermessage, "!addcom") || strings.Contains(usermessage, "!setperm") || strings.Contains(usermessage, "!delcom") {
-					com = CommandOperations(usermessage)
-				}
-
-				if strings.Contains(usermessage, "!edittimed") || strings.Contains(usermessage, "!addtimed") {
-					TimedCommandOperations(usermessage)
-				}
-
-				if strings.Contains(usermessage, "!settitle") {
-					changeTitleSplit := strings.Split(usermessage, " ")
-					PostStreamData(irc, irc.conn, irc.ChannelName, "title", changeTitleSplit[1:])
-				}
-
-				if strings.Contains(usermessage, "!setgame") {
-					changeGameSplit := strings.Split(usermessage, " ")
-					PostStreamData(irc, irc.conn, irc.ChannelName, "game", changeGameSplit[1:])
-				}
-
-				if usermessage == "!startraffle" {
-					//raffleRunning = true
-					go GameRoot(irc, username, usermessage, "raffle")
-					BotSendMsg(irc.conn, irc.ChannelName, "A points raffle has just started. Type !raffle <amount> to enter the raffle for a chance to score big!", irc.BotName)
-				}
-
-				if strings.Contains(usermessage, "!newgiveaway") {
-					giveawaySplit := strings.Split(usermessage, " ")
-					giveawayEntryTerm = giveawaySplit[1]
-
-					BotSendMsg(irc.conn, irc.ChannelName, "A new giveaway has started! Type '"+giveawayEntryTerm+"' to enter!", irc.BotName)
-
-				}
-
-				if strings.Contains(usermessage, "!givepoints") {
-					splitMessage := strings.Split(usermessage, " ")
-					pointsToGive, _ := strconv.Atoi(splitMessage[2])
-					GivePoints(database, username, pointsToGive)
-				}
-
-				if strings.Contains(usermessage, "!endgiveaway") {
-					if giveawayEntryTerm != "giveawayisnil" {
-						rand.Seed(time.Now().Unix())
-						winner := giveawayUsers[rand.Intn(len(giveawayUsers))]
-						giveawayEntryTerm = "giveawayisnil"
-
-						giveawayUsers = giveawayUsers[:0]
-						BotSendMsg(irc.conn, irc.ChannelName, winner+" is the winner!", irc.BotName)
-					} else {
-						BotSendMsg(irc.conn, irc.ChannelName, "There is no giveaway running.", irc.BotName)
-					}
-				}
-
-				if strings.Contains(usermessage, "!caster") {
-					casterSplit := strings.Split(usermessage, " ")
-					casterTargetMessage := strings.Replace(irc.CasterMessage, "{target}", casterSplit[1], -1)
-					BotSendMsg(irc.conn, irc.ChannelName, casterTargetMessage, irc.BotName)
-				}
-
-				if usermessage == "!listcoms" {
-					paste := PostPasteBin(irc.PastebinKey, com)
-					BotSendMsg(irc.conn, irc.ChannelName, "Command list: "+paste, irc.BotName)
-
-				}
-
-				if strings.Contains(usermessage, "!permit") {
-					permitSplit := strings.Split(usermessage, " ")
-					permUsers = append(permUsers, permitSplit[1])
-					BotSendMsg(irc.conn, irc.ChannelName, permitSplit[1]+" can now post one link in chat.", irc.BotName)
-				}
-			}
-			if usermessage == giveawayEntryTerm {
-				giveawayUsers = append(giveawayUsers, username)
-				fmt.Println(giveawayUsers)
-			}
-			if usermessage == "!"+irc.PointsName {
-				userPoints := GetUserPoints(username)
-				pointString := strconv.Itoa(userPoints)
-				pointsTargetMessage := strings.Replace(irc.PointsMessage, "{target}", username, -1)
-				pointsTargetMessage = strings.Replace(pointsTargetMessage, "{value}", pointString, -1)
-				pointsTargetMessage = ReplaceStrings(pointsTargetMessage, "{currency}", irc.PointsName)
-				BotSendMsg(irc.conn, irc.ChannelName, pointsTargetMessage, irc.BotName)
+				com, quotes, goofs.GoofSlice = CreateCommands(irc, com, quotes, badwords, goofs, usermessage, database, line)
 			}
 
-			if strings.Contains(usermessage, "!raffle") {
-				participating := make(map[string]chan int)
-				go RafflePoints(irc, username, usermessage, participating)
-				newUser := <-participating[username]
-				fmt.Println(newUser)
-			}
-
-			if usermessage == "!game" {
-				game := GetGame(irc.conn, irc.ChannelName)
-				var gameName string
-				if len(game.Data) > 0 {
-					for _, val := range game.Data {
-						gameName = val.Name
-						BotSendMsg(irc.conn, irc.ChannelName, "@"+username+", "+gameName, irc.BotName)
-					}
-				} else {
-					BotSendMsg(irc.conn, irc.ChannelName, "@"+username+", stream is offline.", irc.BotName)
-				}
-			}
-
-			if irc.GamesEnabled == true {
-				if strings.Contains(usermessage, "!roulette") {
-					go GameRoot(irc, username, usermessage, "roulette")
-				}
-			}
+			DefaultCommands(irc, username, usermessage, line, com, quotes, badwords, goofs, permUsers, giveawayEntryTerm, giveawayUsers, database)
 
 			// Check for occurences of values from arrays/slices/maps etc
 
@@ -645,71 +497,10 @@ func main() {
 				}
 			}
 
-			for k, v := range quotes {
-				if usermessage == "!quote "+k {
-					BotSendMsg(irc.conn, irc.ChannelName, v, irc.BotName)
-				}
-
-			}
-			if usermessage == "!quote" {
-				rows, err := database.Query("SELECT QuoteID, QuoteContent from quotes ORDER BY RANDOM() LIMIT 1;")
-				if err != nil {
-					fmt.Printf("Error: %s", err)
-				}
-				for rows.Next() {
-					var QuoteID string
-					var QuoteContent string
-					rows.Scan(&QuoteID, &QuoteContent)
-					quotes[QuoteID] = QuoteContent
-					BotSendMsg(irc.conn, irc.ChannelName, QuoteContent, irc.BotName)
-				}
-			}
-
-			// Check if user typed in !addgoof in the chat
-			checkForGoof := strings.Contains(usermessage, "!addgoof")
-			if checkForGoof == true {
-				// Split data to separate username from value to use as new goof
-				GoofSplit := strings.Split(usermessage, " ")
-				GoofString := string(GoofSplit[1])
-				fmt.Println(GoofSplit[1])
-
-				statement, err := database.Prepare("INSERT INTO goofs (GoofName) VALUES (?)")
-				if err != nil {
-					fmt.Printf("Error: %s", err)
-				}
-				statement.Exec(GoofString) // Inserts value of GoofString into the (?) in previous SQL statement
-
-				// Append to the slice in this run session to make it useable right away
-				goofs.GoofSlice = append(goofs.GoofSlice, GoofString)
-
-			}
-
-			// Check if usermessage has !addquote in it
-			CheckForAddQuote := strings.Contains(usermessage, "!addquote")
-			if CheckForAddQuote == true {
-				// Check if user is moderator or broadcaster
-				if CheckUserStatus(line, "moderator", irc) == "true" {
-					quotes = AddQuote(irc.conn, irc.ChannelName, line, usermessage, irc.BotName)
-				} else if CheckUserStatus(line, "broadcaster", irc) == "true" {
-					quotes = AddQuote(irc.conn, irc.ChannelName, line, usermessage, irc.BotName)
-				} else {
-					BotSendMsg(irc.conn, irc.ChannelName, "Must be a moderator to add a new quote.", irc.BotName)
-				}
-
-			}
-
-			// Respond to user the current time, currently locked to the computer the bot is running on
-			if usermessage == "!time" {
-				if irc.StreamerTimeToggle == true {
-					TimeCommands("StreamerTime", irc.conn, irc.ChannelName, irc.BotName, username)
-				}
-			}
-
-			if usermessage == "!uptime" {
-				TimeCommands("Uptime", irc.conn, irc.ChannelName, irc.BotName, username)
-			}
-
 		} else if strings.Contains(line, "USERNOTICE") {
+			currenttime := time.Now()
+			datestring := currenttime.String()
+			datesplit := strings.Split(datestring, " ")
 			// user variables used to split the twitch tag string to get the username
 			if strings.Contains(line, "msg-param-sub-plan") {
 				if irc.RespondToSubs == true {
